Remove duplicated with_request_store spec-metadata
The `with_request_store` metadata already lived inside `spec_helper`
so we don't need another one.

This also replaces the `before` and `after` blocks for handling the
request store with a single around block. The advantage of that is
that this `around` block would be wrapped around any `around` blocks
defined inside the specs. Which I think is more predictable behaviour.