Use an existing image for testing /help
In spec/controllers/help_controller_spec.rb we test how an image is
rendered under /help. Given that we usually rename images as we
update them, we should use one that remains unchanged.

https://gitlab.com/gitlab-org/gitlab/-/merge_requests/24139 fixed
a failure and introduced a new image. Let's use one that we already
use in the docs and we're unlikely to change its name.