Fix display logic of Securty Report MR widget
The Security Report MR widget was being displayed if _only_ the license
compliance CI job was set up for the project, even though it would not
have any data to display for its report types.

This tightens up that logic, so that it is only displayed when one or
more of the security reports is set up for the project.

Fixes https://gitlab.com/gitlab-org/gitlab/issues/200017.