Support Workhorse directly uploading files to S3
This supports the AWS S3 client that will be used by Workhorse in
https://gitlab.com/gitlab-org/gitlab-workhorse/-/merge_requests/466.

This makes it possible to use S3 buckets with default KMS encryption and
proper MD5 checksums.

The Workhorse S3 client is only enabled for instance profiles and V4
signatures.

Since instance profiles are an AWS-specific feature, we should be
reasonably confident that object storage will work with the Workhorse
AWS S3 client.