Fix Service Side Request Forgery in JenkinsDeprecatedService
We noticed requests on some ReactiveCache jobs going to localhost, and
it appears that JenkinsDeprecatedService may have had legacy code to
allow localhost requests instead of using the setting defaults. This
likely happened because all the changes were made in CE and not EE.

To fix this, we now remove the explicit argument and allow `UrlBlocker`
to use the default settings.