Fix broken external links
As external sites update their architecture, things get redirected,
deleted, etc... and we want to make sure we don't have dead links in
the docs. At the same time, redirected links aren't necessarily bad,
but we want to make sure we aren't redirecting to a page that no
longer lists the information, so it's good to update the redirected
links as well.