Application limit for ES indexed field length
Using the plan_limit table we can down store a limit for the maximum
size of any fields being indexed in Elasticsearch. This is for
https://gitlab.com/gitlab-org/gitlab/issues/201826 . Any strings above
the length limit will be truncated down to the limit.

We default to 0 which means unlimited so this has no affect for
self-managed customers. On GitLab.com we will likely want to set this to
around 20k as a sensible upper limit on what we should store.