Update feature_categories.yml with new data
We've added some new categories on our website. The only worker I could
find with a clear change of category was the
RemoveUnreferencedLfsObjectsWorker, which is now part of git_lfs rather
than source_code_management.

For other workers, we might now be wrong. However, the idea behind the
attribution is that if the relevant group thinks a worker is attributed
to them incorrectly, they can go fix it - it's not essential that these
are 100% accurate all the time.