Standardize project fabrication
In some places on QA tests in the Plan stage, when creating a
project, we do:

```ruby
Resource::Project.fabricate_via_api! do |resource|
  resource.name = 'project-name'
end
```

But in the majority of the tests, we do:

```ruby
Resource::Project.fabricate_via_api! do |project|
  project.name = 'project-name'
end
```

So better to stick with one way of doing it for standardization.