Avoid double encoding of import url credentials
Import projects by URL, when the URLs credentials have characters that
require percent encoding, is currently not working. This feature is
failing because the import url credentials are being encoded before
saving on the database, on the `Project#import_url=`, and then it's
being encoded again when retrieving the credentials from the database,
on `Project#import_url`.

To solve the problem the credential is being percent decoded before
saving it on the database.