Add unix to allowed hosts
In Rails 6, the new HostAuthorization middleware is enabled by
default (it's not enabled on other environments). It currently
allows localhost and IP addresses and we add
`Gitlab.config.gitlab.host` as additional allowed host.

When gitlab-shell makes a request via unix socket, it uses
`http://unix/<endpoint>` so the host will be `unix`. Since it's
not included in the list of allowed hosts, it gets blocked.

This adds `unix` host to the list of allowed hosts.