Automatically detect Gitaly feature flags
For Gitaly developers, Rails was a required codebase to change when
working with feature flags. That was great for our throughput numbers,
but not a great change to make. With this change, each Gitaly feature
flag will be detected and send to the Gitaly server by default.

Given this change relies on a full set of the persisted feature flags,
the potentially stale cache is used. This reduces load on the database
server, and given on GitLab.com has many nodes this would still give the
illusion of starting to use the feature flag right away. For our
customers the impact is larger, as they might need to wait upto a
minute, though on average 30s, for a feature flag to take effect.