Make elasticsearch bulk parameters configurable
These variables strongly affect the amount of load that indexing
operations place on the Elasticsearch server. They are useful values
to be able to twiddle in both performance and availability contexts.

At present, these only affect `gitlab-elasticsearch-indexer` behaviour,
but we can roll it out to the sidekiq context over time.