Expire the Gitlab::CurrentSettings cache
Expire the cache after it the self_monitoring_project_id is saved. This
is for the benefit of the later step where the Prometheus settings are
saved. When the Prometheus settings are saved, the PrometheusService
needs to be able to read the self_monitoring_project_id. If the cache
is not expired, PrometheusService will see the old value of the
self_monitoring_project_id. There is code in PrometheusService that
allows the self_monitoring_project to connect to the internal Prometheus
without needing local URLs to be allowed in webhooks.