Fix time-dependent spec failures in user_sees_error_details_spec.rb
These features specs were failing because they were using fixed dates
and attempting to use Timecop, but Timecop doesn't apply to the browser.
When the date crossed the 8th of the month, the Timeago text "1 month
ago" advanced to "2 months ago".

To fix this issue, we now adjust the test data to use a timestamp
relative to today's date, eliminating the need for Timecop.

Closes https://gitlab.com/gitlab-org/gitlab/issues/202678