Upgrade indirect dependency caniuse-lite
Upgrade 'caniuse-lite' indirect dependency.

When the 'caniuse-lite' dependency is more than 6 months old, it
starts showing a warning that it is outdated.

This related MR
(https://gitlab.com/gitlab-org/gitlab/-/merge_requests/23030)
avoids this in some cases, but not others (e.g. RubyMine
linting integration).

To avoid the warning in all cases, we need to update it every
six months.

But, it's an indirect dependency, so the upgrade can't be done via a
straightforward 'yarn upgrade'

Instead, in order upgrade it, you must use this approach
(https://github.com/yarnpkg/yarn/issues/4986):

1. Delete the 'caniuse-lite' entry from
2. Re-run 'yarn install'