Replace browser confirm with GlModal (Geo)
Based on our new design specs we want to replace default browser
behavior with our Pajamas Design System.

This required us to replace the confirm with a modal confirm.
GlModal information can be found here:
https://gitlab-org.gitlab.io/gitlab-ui/?path=/story/base-modal--default

This was not currently feasible so this work is built on top of
a new JS Confirm Modal that will be pulled into its own MR.
The issue tracking that info can be found here:
https://gitlab.com/gitlab-org/gitlab/issues/205900