Remove reference to outdated issue in comments
This commit removes a reference to an outdated issue from comments
within vuex-action files.

It pointed to an issue that aimed to make the default export obsolete,
but the issue has been closed and the export will be needed until
all tests have been migrated to karma.

See comment on the issue:
https://gitlab.com/gitlab-org/gitlab/issues/24511#note_273343173