Add comment with reason for skipping admin auth
Updating Doorkeeper to 5.0.2 introduced a change where all controllers
that inherit from `Doorkeeper::ApplicationsController` must define the
admin_authenticator in the config, otherwise all the endpoints will
always return a `403`.

As we have our own authenication logic for this controller defined in
`authenticate_user!`, this call is redundant. We've chosen to explicitly
skip it here, so that anyone who inherits from this controller in the
future will have to decide how to handle the authenication explicitly.