Rework queue selector operators
We now have `|` for OR, `&` for AND, and `,` for concatenating
sets. This is hopefully more familiar to programmers and SREs.

Additionally, update the documentation to clarify that `=` and `!=` work
on the sets to their right, so they're more like 'in' and 'not in' than
'equal to' and 'not equal to'.