Fix autocomplete limitation bug
jQuery at.js has a hard limit on how many items can be searched. Before
it executes it's search function, it limits the array size to 1000
items. We have been hitting this bug because we have more than a
thousand users or labels [0][1].

We are planning to replace at.js [2], but in order to fix this bug for
now, we have forked at.js [3] and made adjustments, so that it works for
our usecases.

Performance wise this should have no big impact, as we are rendering a
maximum of 5 results in the most cases anyway. The search on the
javascript side is rather performant.

[0]: https://gitlab.com/gitlab-org/gitlab/issues/204825
[1]: https://gitlab.com/gitlab-org/gitlab/issues/198650
[2]: https://gitlab.com/gitlab-org/gitlab/issues/27240
[3]: https://gitlab.com/gitlab-org/frontend/At.js