Omit previous error from Sidekiq JSON logs
Sidekiq stores a job's error details in the payload for the _next_ run,
so that it can display the error in the Sidekiq UI. This is because
Sidekiq's main state is the queue of jobs to be run. However, in our
logs, this is very confusing, because we shouldn't have any error at all
when a job starts, and we already add an error message and class to our
logs when a job fails.