Add QA test of fork creation after storage change
Tests that if a fork is created and then the parent project is moved
to a new gitaly storage/shard, another fork can be created after the
move.

See https://gitlab.com/gitlab-org/gitlab/issues/38264

Add the ability to create a fork via the API, check import status (to
confirm when the fork is complete), and change repository storage.

Resource::Fork needed to be updated, which also required updating
Resource::MergeRequestFromFork and the test that uses it so that that
test can still fabricate a fork via the UI.