Count threads all that run in multithreaded env
This adds the threads we know are running next to the worker threads
for Puma and Sidekiq. Adding them here makes sure our connection pool
for the database is large enough.

In puma, the main thread also connects to the database, so we need to
count that.

In Sidekiq, the main thread and the poller for Sidekiq cron could also
connect to the database.