Fix dismissed vulnerability comment updates
Context: We currently use two different data-sources for
displaying dismissal-feedback inside the vulnerability modal.

Within the MR-widget the data from the security-report is used
before the response from the vulnerability-feedback-endpoint.

The issue is that the security report endpoints are being cached,
so any changes to the dismissal comments are not reflected in the UI
until the cache is stale.

By using the data from the dedicated feedback-endpoint this can be
avoided.