Documentation: external_url and internal_api_url configuration
If you blindly follow the steps on setting Prefect support for GitLab
it is hard to notice that 'internal_api_url' value for Gitaly
configuration needs to be changed to the proper one. So this commit
fixes it by adding explicit instructions about how it needs to be
changed. Also this commit adds 'external_url' configuration for GitLab
configuration file and comment on how it needs to be setup.