Update new snippet elements
The UI was changed so we need to update the page object.
Part of the UI is rendered using `_zen.html.haml` based on
parameters defined in `_form.html.haml`, so the element
`description_field` is actually defined by the selector in
`_form.html.haml` AND the code added to `_zen.html.haml`.
So we needed a fake 'element' to allow the sanity check
to verify that the code added to `_zen.html.haml` isn't changed

Removes the test from quarantine because the cause was clear
and the failures were consistent.

See https://gitlab.com/gitlab-org/gitlab/issues/205511