Fix a log schema inconsistency
In JiraService, some structured logs for errors were sent with a Hash in
the error key, and some with a String. This frustrates Elasticsearch,
which is commonly used to store structured logs, causing it to have to
drop logs from one instance or the other due to the type mismatch.