Fix a log schema inconsistency
In most of the application, structured logs that emit an "exception" key
for errors are sent with a Hash in this key. In this class, a string was
used. This frustrates Elasticsearch, which is commonly used to store
structured logs, causing it to have to drop logs from one instance or
the other due to the type mismatch.