Version pin postgres service
We guide the users to use `latest` which is dangerous since it can lead
to breaking changes for things we don't have control over. Specify the
latest version of postgres as an example, and show that the password
needs to be set from now on since postgres updated their image
configuration to always require the password.

reference https://gitlab.com/gitlab-org/gitlab-runner/issues/8323