Remove `role` field and add `job_title` field
The `role` field is removed from the edit profile page and is replaced
with a `job_title` field. This new `job_title` field is displayed on
the profile page and in the user popover. If `job_title` and
`organization` are both set they are displayed as
"#{job_title} at ${organization}"