Add Board Milestones to Group export
We want to be able to export the milestones on Boards for groups as
well.

Prior to this change, adding `:milestone` to the export config file
would result in an error when the `serialize_hash` method was called on
the MilestoneStruct, which is used for predefined, system wide
milestones.

Add this method to the struct so that it responds to the needed
interface and add `:milestone` to the config file for export.