Ignore predefined relations on Import
Currently our Milestone model has some predefined instances that are not
actually persisted to the database (only their ID is persisted on the
related model).

We don't want to import these as actual relations in the database as
their predefined IDs imply some additional functionality in the system -
which is lost if they are imported as a new record.

For now, ignore these records on import as we do not have a clear way to
handle them - but we still need to import the user generated milestones.