Rename methods for better clarity on purpose they serve
* Renamed Note#cross_reference to Note#system_note_with_references? to
reflect that it checks for system notes that may have references
regardless if references are cross project, group or not

* Renamed Note#cross_reference_visible_for?(user) to
Note#system_note_with_references_visible_for?(user) for clarity.
It should be easier to understand that the methods checks if system
notes have referenes and those references are  visible/readable
by prvided user.

* Renamed Epic#cross_reference?(from) to Epic#cross_referenced?(from)
as it does not check if `from` is a cross reference, but rather
if `from` is different from epic's group, which means epic is being
cross referenced.

* Renamed IssuableStateFilter#cross_reference?(issuable) to
IssuableStateFilter#cross_referenced?(issuable) for same reasons as
for Epic