Update GitLab Import to log to all exceptions
We want to ensure that all exceptions go through the same exception
handler so that we have consistent error handling.

Update manual logging instances to use the shared handler so that we can
take advantage of the exception handling introduced in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/17819

Ensure that all raised errors are logged to this shared handler before
continuing to raise.

Addresses https://gitlab.com/gitlab-org/gitlab/issues/38101