Add a Cluster Metric Embed Banzai Filter
Part of the work to, embed cluster health metrics charts in any
Notable (comment, issue description) via adding a url to the gitlab
chart in a field that supports Gilab Flafored Markdown(GFM).
This MR adds a div with a distinctive class name related to metrics to
a Gitlab Flavored Markdown field. The field is redacted by default based
on its class name because permissions are not yet added to
permissions_by_route in this MR.