Add temporary snapshot tests
This adds snapshot tests for the `vulnerability-details` component as it
currently exists in the two main contexts it appears in:

- in Merge Requests
- any Security Dashboard

Snapshots for both contexts are created, because the data is processed
in slightly (though very similar) ways in each.

A relatively wide selection of real vulnerability findings is used in
both contexts to ensure good coverage.

The purpose is to establish a baseline to facilitate refactoring the
component, and the data paths in both contexts, without changing the
final rendered output.