Fix upgrade failure in EE displaying license
When upgrading to 12.8 EE, the initializer might fail with
`NoMethodError: undefined method `pluralize' for LicenseHelper:Module`
if the license neared expiry. This happened
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/24998 changed the
database check so that the license warning would actually be displayed,
but inside an initializer `pluralize` is not available.

The easiest way to fix this now is to revert the database check to the
original version.

Closes https://gitlab.com/gitlab-org/gitlab/issues/207737