Add Strategy Component for Feature Flags
As part of the redesign of feature flags to allow a single strategy to
apply to multiple environments, a new component that acts as a form for
a strategy is introduced.

Strategy - Handles setting the strategy type, as well as any necessary
parameters to go along with that strategy.

New Environments Dropdown - A new dropdown that searches and
autocompletes on any environment scopes for feature flags, very similar
to the currently existing Environments Dropdown.