Hook up API and Validation to Node Form
This change is broken down from a very large MR:
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/22719

The overall goal is to convert the Geo Node Form from HAML to Vue.
Based on MVC, splitting this into smaller MRs is more feasible.

This MR adds the final pice to convert the Geo Form into Vue:
  - The API calls to create/update
  - Form Validation

After this MR is merged, we can do a follow up MR to:
  - Delete legacy code
  - Remove Feature Flag