Use CI source project for path to exposed job artifacts
When routing paths to exposed job artifacts, we were always using the
destination project. However, in a cross-project MR, the job exposing
the artifacts is run in the source project, not in the destination
project. Change the arguments to project_job_path() to always use the
job's project for constructing the URL path.

!23738 tried a more expansive fix but this wasn't doable thanks to
project_job_path() also being a hidden Rails-generated helper function
which did take both project and job as an argument.

Fixes: #37341