Delete data violating the new unique index
It is possible that there is data violating the
new unique index on type and projcet_id
on the services table.

This commit adds a SQL statement to the migration
which deletes those ghost services. On the
application level we use `first_or_create` to find
a project service of a type. This will return the
services with the lowest ID. The migration will
delete services with higher IDs to avoid deleting
data that is actually in use.