API endpoint to remove jobs from Sidekiq based on metadata
This API endpoint uses job metadata to remove jobs from a queue. It can
only be used by admins, and deletes as many jobs as it can in 30
seconds. If it exceeds 30 seconds, it returns a flag indicating that the
user should try again to finish processing the queue.

(Because of the way Sidekiq queues work, it can't resume where it left
off, so you just have to start from scratch each time.)

By default, it's implemented in GraphQL, but there's also a matching
REST endpoint for convenience.