Fix N+1 in milestone show page
The issue is that using becomes means a new object is instantiated all
the time, which means we throw away the route preload for the namespace

In fact, we don't even need to pass in the namespace for the polymorphic
route, the issuable_path helper seems to be able to figure out the
correct path just from the project.

Adds test for n+1 for group milestone show page. We are loading the
route for the namespace for each issue for the milestone.

Also, use the same link for both issue links

Also, removes @project_namespace definition as nothing use it anymore