Remove sub_group_webhooks feature falg
The feature flag was introduced because there
was no limit for the number of webhooks in a group.
This raised a concern that enabling group hooks to
be triggered from a subgroup could result in a large
number of requests per action.

With https://gitlab.com/gitlab-org/gitlab/-/merge_requests/25129
we introduced a limit for webhooks on a group
and the feature flag is not needed anymore