Refactor resizeable-panel from CollapsibleSidebar
This is an incremental refactoring to support a collapsible left sidebar
in the Web IDE.

We want to make the resizeable panel not need to be the responsibility
of CollapsibleSidebar anymore.

This restructure required some styling changes across elements,
especially due to the mix of Flexbox and non-Flexbox in some child
pane components.

It also contains some minor fixes:

* related to removing tooltips and focus highlighting when clicking
  on the buttons.

* Removing header slots which will not be needed by the left side
  and leaving single remaining slot for the footer

* Remove unnecessary h-100 w-100 and nested elements

* Change initial width of resizable-panel in right.vue

* Move classes set in right.vue into resizable-panel since
  it seems like the component should manage this responsibility.

* Clean up some spacing issues in the Web Terminal tab

* Add overflow-hidden class to js-tab-view element since
  this will effect how flex sizes it's children.