Memoize loading of CI variables
When creating a pipeline with many builds, we load group and project
variables repeatedly, which can cause a high number of Redis calls,
database queries, and RAM usage. We can cache this data within
RequestStore to speed up the building of pipelines.

Relates to https://gitlab.com/gitlab-org/gitlab/issues/198694