Disable Marginalia line backtrace in production
As mentioned in https://github.com/basecamp/marginalia/pull/93/files,
adding `:line` to the output of the Marginalia incurs some overhead
because a regexp on the backtrace has to be run on every SQL query. Only
enable this in development because we've seen it slow things down.

Identified in https://gitlab.com/gitlab-org/gitlab/issues/198694